Code/app/requirements.txt
numpy
streamlit
streamlit_option_menu
seaborn
matplotlib
plotly
pandas
joblib
scikit-learn
reportlab
pillow
fpdf2
pdfkit
wkhtmltopdf
weasyprint
chardet
streamlit-aggrid

